Use shorter date format

+++ js/retrieve-posts.js
@@ -2,7 +2,7 @@ $(function() {
     /*$.get("https://api.npoint.io/9f01f9e5f6462ffc8e5e", function(allPosts) {*/
     $.get("json/posts.json", function(allPosts) {
         let postsArea = $('.posts-area');
-        let dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
+        let dateOptions = { weekday: 'short', year: 'numeric', month: 'short', day: 'numeric' };
 
         for (post of allPosts) {
